🚚 api: move tests to proper context
