+++ api/tests/integration/domain/services/certification-challenges-service_test.js
