+++ api/tests/shared/unit/domain/services/get-translated-text_test.js
@@ -1,5 +1,5 @@
-import * as service from '../../../../src/shared/domain/services/get-translated-text.js';
-import { expect } from '../../../test-helper.js';
+import * as service from '../../../../../src/shared/domain/services/get-translated-text.js';
+import { expect } from '../../../../test-helper.js';
 describe('Unit | Domain | Services | get-translated-text', function () {
   describe('#getTranslatedKey', function () {
     const translatedKey = {
