+++ api/tests/shared/unit/infrastructure/adapters/answer-status-database-adapter_test.js
@@ -1,6 +1,6 @@
-import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
-import * as AnswerStatusDatabaseAdapter from '../../../../src/shared/infrastructure/adapters/answer-status-database-adapter.js';
-import { expect } from '../../../test-helper.js';
+import { AnswerStatus } from '../../../../../src/shared/domain/models/index.js';
+import * as AnswerStatusDatabaseAdapter from '../../../../../src/shared/infrastructure/adapters/answer-status-database-adapter.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('AnswerStatusDatabaseAdapter', function () {
   describe('#adapt', function () {
